+++ x/accounts/defaults/lockup/delayed_locking_account.go
@@ -6,7 +6,7 @@ import (
 
 	"cosmossdk.io/math"
 	"cosmossdk.io/x/accounts/accountstd"
-	lockuptypes "cosmossdk.io/x/accounts/lockup/types"
+	lockuptypes "cosmossdk.io/x/accounts/defaults/lockup/types"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
@@ -131,7 +131,7 @@ func (dva DelayedLockingAccount) QueryVestingAccountInfo(ctx context.Context, re
 	if err != nil {
 		return nil, err
 	}
-	hs := dva.headerService.GetHeaderInfo(ctx)
+	hs := dva.headerService.HeaderInfo(ctx)
 	unlockedCoins, lockedCoins, err := dva.GetLockCoinsInfo(ctx, hs.Time)
 	if err != nil {
 		return nil, err