+++ x/gov/keeper/abci.go
@@ -11,7 +11,6 @@ import (
 	"cosmossdk.io/collections"
 	"cosmossdk.io/core/event"
 	"cosmossdk.io/core/router"
-	"cosmossdk.io/log"
 	"cosmossdk.io/x/gov/types"
 	v1 "cosmossdk.io/x/gov/types/v1"
 
@@ -21,41 +20,50 @@ import (
 
 // EndBlocker is called every block.
 func (k Keeper) EndBlocker(ctx context.Context) error {
-	defer telemetry.ModuleMeasureSince(types.ModuleName, time.Now(), telemetry.MetricKeyEndBlocker)
+	defer telemetry.ModuleMeasureSince(types.ModuleName, telemetry.Now(), telemetry.MetricKeyEndBlocker)
 
-	logger := k.Logger()
 	// delete dead proposals from store and returns theirs deposits.
 	// A proposal is dead when it's inactive and didn't get enough deposit on time to get into voting phase.
-	rng := collections.NewPrefixUntilPairRange[time.Time, uint64](k.environment.HeaderService.GetHeaderInfo(ctx).Time)
-	err := k.InactiveProposalsQueue.Walk(ctx, rng, func(key collections.Pair[time.Time, uint64], _ uint64) (bool, error) {
-		proposal, err := k.Proposals.Get(ctx, key.K2())
+	rng := collections.NewPrefixUntilPairRange[time.Time, uint64](k.HeaderService.HeaderInfo(ctx).Time)
+	iter, err := k.InactiveProposalsQueue.Iterate(ctx, rng)
+	if err != nil {
+		return err
+	}
+
+	inactiveProps, err := iter.KeyValues()
+	if err != nil {
+		return err
+	}
+
+	for _, prop := range inactiveProps {
+		proposal, err := k.Proposals.Get(ctx, prop.Key.K2())
 		if err != nil {
 			// if the proposal has an encoding error, this means it cannot be processed by x/gov
 			// this could be due to some types missing their registration
 			// instead of returning an error (i.e, halting the chain), we fail the proposal
 			if errors.Is(err, collections.ErrEncoding) {
-				proposal.Id = key.K2()
-				if err := failUnsupportedProposal(logger, ctx, k, proposal, err.Error(), false); err != nil {
-					return false, err
+				proposal.Id = prop.Key.K2()
+				if err := failUnsupportedProposal(ctx, k, proposal, err.Error(), false); err != nil {
+					return err
 				}
 
 				if err = k.DeleteProposal(ctx, proposal.Id); err != nil {
-					return false, err
+					return err
 				}
 
-				return false, nil
+				continue
 			}
 
-			return false, err
+			return err
 		}
 
 		if err = k.DeleteProposal(ctx, proposal.Id); err != nil {
-			return false, err
+			return err
 		}
 
 		params, err := k.Params.Get(ctx)
 		if err != nil {
-			return false, err
+			return err
 		}
 		if !params.BurnProposalDepositPrevote {
 			err = k.RefundAndDeleteDeposits(ctx, proposal.Id) // refund deposit if proposal got removed without getting 100% of the proposal
@@ -64,26 +72,26 @@ func (k Keeper) EndBlocker(ctx context.Context) error {
 		}
 
 		if err != nil {
-			return false, err
+			return err
 		}
 
 		// called when proposal become inactive
 		// call hook when proposal become inactive
-		if err := k.environment.BranchService.Execute(ctx, func(ctx context.Context) error {
+		if err := k.BranchService.Execute(ctx, func(ctx context.Context) error {
 			return k.Hooks().AfterProposalFailedMinDeposit(ctx, proposal.Id)
 		}); err != nil {
 			// purposely ignoring the error here not to halt the chain if the hook fails
-			logger.Error("failed to execute AfterProposalFailedMinDeposit hook", "error", err)
+			k.Logger.Error("failed to execute AfterProposalFailedMinDeposit hook", "error", err)
 		}
 
-		if err := k.environment.EventService.EventManager(ctx).EmitKV(types.EventTypeInactiveProposal,
+		if err := k.EventService.EventManager(ctx).EmitKV(types.EventTypeInactiveProposal,
 			event.NewAttribute(types.AttributeKeyProposalID, fmt.Sprintf("%d", proposal.Id)),
 			event.NewAttribute(types.AttributeKeyProposalResult, types.AttributeValueProposalDropped),
 		); err != nil {
-			logger.Error("failed to emit event", "error", err)
+			k.Logger.Error("failed to emit event", "error", err)
 		}
 
-		logger.Info(
+		k.Logger.Info(
 			"proposal did not meet minimum deposit; deleted",
 			"proposal", proposal.Id,
 			"proposal_type", proposal.ProposalType,
@@ -91,42 +99,49 @@ func (k Keeper) EndBlocker(ctx context.Context) error {
 			"min_deposit", sdk.NewCoins(proposal.GetMinDepositFromParams(params)...).String(),
 			"total_deposit", sdk.NewCoins(proposal.TotalDeposit...).String(),
 		)
+	}
 
-		return false, nil
-	})
+	// fetch active proposals whose voting periods have ended (are passed the block time)
+	rng = collections.NewPrefixUntilPairRange[time.Time, uint64](k.HeaderService.HeaderInfo(ctx).Time)
+
+	iter, err = k.ActiveProposalsQueue.Iterate(ctx, rng)
 	if err != nil {
 		return err
 	}
 
-	// fetch active proposals whose voting periods have ended (are passed the block time)
-	rng = collections.NewPrefixUntilPairRange[time.Time, uint64](k.environment.HeaderService.GetHeaderInfo(ctx).Time)
-	err = k.ActiveProposalsQueue.Walk(ctx, rng, func(key collections.Pair[time.Time, uint64], _ uint64) (bool, error) {
-		proposal, err := k.Proposals.Get(ctx, key.K2())
+	activeProps, err := iter.KeyValues()
+	if err != nil {
+		return err
+	}
+
+	// err = k.ActiveProposalsQueue.Walk(ctx, rng, func(key collections.Pair[time.Time, uint64], _ uint64) (bool, error) {
+	for _, prop := range activeProps {
+		proposal, err := k.Proposals.Get(ctx, prop.Key.K2())
 		if err != nil {
 			// if the proposal has an encoding error, this means it cannot be processed by x/gov
 			// this could be due to some types missing their registration
 			// instead of returning an error (i.e, halting the chain), we fail the proposal
 			if errors.Is(err, collections.ErrEncoding) {
-				proposal.Id = key.K2()
-				if err := failUnsupportedProposal(logger, ctx, k, proposal, err.Error(), true); err != nil {
-					return false, err
+				proposal.Id = prop.Key.K2()
+				if err := failUnsupportedProposal(ctx, k, proposal, err.Error(), true); err != nil {
+					return err
 				}
 
 				if err = k.ActiveProposalsQueue.Remove(ctx, collections.Join(*proposal.VotingEndTime, proposal.Id)); err != nil {
-					return false, err
+					return err
 				}
 
-				return false, nil
+				continue
 			}
 
-			return false, err
+			return err
 		}
 
 		var tagValue, logMsg string
 
 		passes, burnDeposits, tallyResults, err := k.Tally(ctx, proposal)
 		if err != nil {
-			return false, err
+			return err
 		}
 
 		// Deposits are always burned if tally said so, regardless of the proposal type.
@@ -142,19 +157,19 @@ func (k Keeper) EndBlocker(ctx context.Context) error {
 			// in case of an error, log it and emit an event
 			// we do not want to halt the chain if the refund/burn fails
 			// as it could happen due to a governance mistake (governance has let a proposal pass that sends gov funds that were from proposal deposits)
-			k.Logger().Error("failed to refund or burn deposits", "error", err)
+			k.Logger.Error("failed to refund or burn deposits", "error", err)
 
-			if err := k.environment.EventService.EventManager(ctx).EmitKV(types.EventTypeProposalDeposit,
+			if err := k.EventService.EventManager(ctx).EmitKV(types.EventTypeProposalDeposit,
 				event.NewAttribute(types.AttributeKeyProposalID, fmt.Sprintf("%d", proposal.Id)),
 				event.NewAttribute(types.AttributeKeyProposalDepositError, "failed to refund or burn deposits"),
 				event.NewAttribute("error", err.Error()),
 			); err != nil {
-				k.Logger().Error("failed to emit event", "error", err)
+				k.Logger.Error("failed to emit event", "error", err)
 			}
 		}
 
 		if err = k.ActiveProposalsQueue.Remove(ctx, collections.Join(*proposal.VotingEndTime, proposal.Id)); err != nil {
-			return false, err
+			return err
 		}
 
 		switch {
@@ -178,10 +193,10 @@ func (k Keeper) EndBlocker(ctx context.Context) error {
 			// Messages may mutate state thus we use a cached context. If one of
 			// the handlers fails, no state mutation is written and the error
 			// message is logged.
-			if err := k.environment.BranchService.Execute(ctx, func(ctx context.Context) error {
+			if err := k.BranchService.Execute(ctx, func(ctx context.Context) error {
 				// execute all messages
 				for idx, msg = range messages {
-					if _, err := safeExecuteHandler(ctx, msg, k.environment.RouterService.MessageRouterService()); err != nil {
+					if _, err := safeExecuteHandler(ctx, msg, k.RouterService.MessageRouterService()); err != nil {
 						// `idx` and `err` are populated with the msg index and error.
 						proposal.Status = v1.StatusFailed
 						proposal.FailedReason = err.Error()
@@ -210,14 +225,14 @@ func (k Keeper) EndBlocker(ctx context.Context) error {
 			proposal.Expedited = false // can be removed as never read but kept for state coherence
 			params, err := k.Params.Get(ctx)
 			if err != nil {
-				return false, err
+				return err
 			}
 			endTime := proposal.VotingStartTime.Add(*params.VotingPeriod)
 			proposal.VotingEndTime = &endTime
 
 			err = k.ActiveProposalsQueue.Set(ctx, collections.Join(*proposal.VotingEndTime, proposal.Id), proposal.Id)
 			if err != nil {
-				return false, err
+				return err
 			}
 
 			if proposal.ProposalType == v1.ProposalType_PROPOSAL_TYPE_EXPEDITED {
@@ -237,18 +252,18 @@ func (k Keeper) EndBlocker(ctx context.Context) error {
 		proposal.FinalTallyResult = &tallyResults
 
 		if err = k.Proposals.Set(ctx, proposal.Id, proposal); err != nil {
-			return false, err
+			return err
 		}
 
 		// call hook when proposal become active
-		if err := k.environment.BranchService.Execute(ctx, func(ctx context.Context) error {
+		if err := k.BranchService.Execute(ctx, func(ctx context.Context) error {
 			return k.Hooks().AfterProposalVotingPeriodEnded(ctx, proposal.Id)
 		}); err != nil {
 			// purposely ignoring the error here not to halt the chain if the hook fails
-			logger.Error("failed to execute AfterProposalVotingPeriodEnded hook", "error", err)
+			k.Logger.Error("failed to execute AfterProposalVotingPeriodEnded hook", "error", err)
 		}
 
-		logger.Info(
+		k.Logger.Info(
 			"proposal tallied",
 			"proposal", proposal.Id,
 			"proposal_type", proposal.ProposalType,
@@ -257,17 +272,15 @@ func (k Keeper) EndBlocker(ctx context.Context) error {
 			"results", logMsg,
 		)
 
-		if err := k.environment.EventService.EventManager(ctx).EmitKV(types.EventTypeActiveProposal,
+		if err := k.EventService.EventManager(ctx).EmitKV(types.EventTypeActiveProposal,
 			event.NewAttribute(types.AttributeKeyProposalID, fmt.Sprintf("%d", proposal.Id)),
 			event.NewAttribute(types.AttributeKeyProposalResult, tagValue),
 			event.NewAttribute(types.AttributeKeyProposalLog, logMsg),
 		); err != nil {
-			logger.Error("failed to emit event", "error", err)
+			k.Logger.Error("failed to emit event", "error", err)
 		}
-
-		return false, nil
-	})
-	return err
+	}
+	return nil
 }
 
 // executes route(msg) and recovers from panic.
@@ -284,7 +297,6 @@ func safeExecuteHandler(ctx context.Context, msg sdk.Msg, router router.Router) 
 
 // failUnsupportedProposal fails a proposal that cannot be processed by gov
 func failUnsupportedProposal(
-	logger log.Logger,
 	ctx context.Context,
 	k Keeper,
 	proposal v1.Proposal,
@@ -308,14 +320,14 @@ func failUnsupportedProposal(
 		eventType = types.EventTypeActiveProposal
 	}
 
-	if err := k.environment.EventService.EventManager(ctx).EmitKV(eventType,
+	if err := k.EventService.EventManager(ctx).EmitKV(eventType,
 		event.NewAttribute(types.AttributeKeyProposalID, fmt.Sprintf("%d", proposal.Id)),
 		event.NewAttribute(types.AttributeKeyProposalResult, types.AttributeValueProposalFailed),
 	); err != nil {
-		logger.Error("failed to emit event", "error", err)
+		k.Logger.Error("failed to emit event", "error", err)
 	}
 
-	logger.Info(
+	k.Logger.Info(
 		"proposal failed to decode; deleted",
 		"proposal", proposal.Id,
 		"proposal_type", proposal.ProposalType,