+++ x/gov/types/v1/msgs_test.go
@@ -10,6 +10,7 @@ import (
 	v1 "cosmossdk.io/x/gov/types/v1"
 
 	"github.com/cosmos/cosmos-sdk/codec"
+	codectestutil "github.com/cosmos/cosmos-sdk/codec/testutil"
 	"github.com/cosmos/cosmos-sdk/codec/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 )
@@ -33,7 +34,9 @@ func init() {
 
 func TestMsgDepositGetSignBytes(t *testing.T) {
 	addr := sdk.AccAddress("addr1")
-	msg := v1.NewMsgDeposit(addr, 0, coinsPos)
+	addrStr, err := codectestutil.CodecOptions{}.GetAddressCodec().BytesToString(addr)
+	require.NoError(t, err)
+	msg := v1.NewMsgDeposit(addrStr, 0, coinsPos)
 	pc := codec.NewProtoCodec(types.NewInterfaceRegistry())
 	res, err := pc.MarshalAminoJSON(msg)
 	require.NoError(t, err)
@@ -44,6 +47,8 @@ func TestMsgDepositGetSignBytes(t *testing.T) {
 // this tests that Amino JSON MsgSubmitProposal.GetSignBytes() still works with Content as Any using the ModuleCdc
 func TestMsgSubmitProposal_GetSignBytes(t *testing.T) {
 	pc := codec.NewProtoCodec(types.NewInterfaceRegistry())
+	addr0Str, err := codectestutil.CodecOptions{}.GetAddressCodec().BytesToString(addrs[0])
+	require.NoError(t, err)
 	testcases := []struct {
 		name         string
 		proposal     []sdk.Msg
@@ -54,7 +59,7 @@ func TestMsgSubmitProposal_GetSignBytes(t *testing.T) {
 	}{
 		{
 			"MsgVote",
-			[]sdk.Msg{v1.NewMsgVote(addrs[0], 1, v1.OptionYes, "")},
+			[]sdk.Msg{v1.NewMsgVote(addr0Str, 1, v1.OptionYes, "")},
 			"gov/MsgVote",
 			"Proposal for a governance vote msg",
 			v1.ProposalType_PROPOSAL_TYPE_STANDARD,